Cast StatFs return values to long
On devices with large storage partitions, we could overflow when
converting free blocks to free bytes. Cast to long to avoid this.

Bug: 3327970

+++ src/com/android/browser/WebStorageSizeManager.java
@@ -141,11 +141,11 @@ public class WebStorageSizeManager {
         }
 
         public long getFreeSpaceSizeBytes() {
-            return mFs.getAvailableBlocks() * mFs.getBlockSize();
+            return (long)(mFs.getAvailableBlocks()) * mFs.getBlockSize();
         }
 
         public long getTotalSizeBytes() {
-            return mFs.getBlockCount() * mFs.getBlockSize();
+            return (long)(mFs.getBlockCount()) * mFs.getBlockSize();
         }
     };
